# docs/swagger.yaml
definitions:
  controllers.AdvertiseInput:
    properties:
      content:
        type: string
      duration:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: string
      user_id:
        type: integer
      vocation_id:
        type: integer
    type: object
  controllers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.PaymentInput:
    properties:
      advertise_id:
        type: integer
      amount:
        type: number
      currency:
        type: string
      due_date:
        type: string
      id:
        type: integer
      user_id:
        type: integer
      vocation_id:
        type: integer
    type: object
  controllers.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.VocationInput:
    properties:
      address:
        type: string
      city:
        type: string
      comment:
        type: string
      country:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      phone:
        type: string
      price:
        type: string
      rating:
        type: string
      user_id:
        type: integer
    type: object
  models.Advertise:
    properties:
      content:
        type: string
      created_at:
        type: string
      duration:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      vocation:
        $ref: '#/definitions/models.Vocation'
      vocation_id:
        type: integer
    type: object
  models.Payment:
    properties:
      advertise:
        $ref: '#/definitions/models.Advertise'
      advertise_id:
        type: integer
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      due_date:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      vocation:
        $ref: '#/definitions/models.Vocation'
      vocation_id:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Vocation:
    properties:
      address:
        type: string
      city:
        type: string
      comment:
        type: string
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      phone:
        type: string
      price:
        type: string
      rating:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /advertise:
    get:
      description: Get a list of Advertise.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Advertise'
            type: array
      summary: Get all Advertise.
      tags:
      - Advertise
    post:
      description: Creating a new Advertise.
      parameters:
      - description: the body to create a new Advertise
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.AdvertiseInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Advertise'
      security:
      - BearerToken: []
      summary: Create New Advertise.
      tags:
      - Advertise
  /advertise-current-user:
    get:
      description: Get All Struct Advertise from Curent User.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Advertise'
            type: array
      security:
      - BearerToken: []
      summary: Get Struct Advertise from Curent User.
      tags:
      - Advertise
  /advertise/{id}:
    delete:
      description: Delete a Advertise by id.
      parameters:
      - description: Advertise id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Advertise.
      tags:
      - Advertise
    get:
      description: Get an Advertise by id.
      parameters:
      - description: Advertise id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Advertise'
      summary: Get Advertise.
      tags:
      - Advertise
    patch:
      description: Update Advertise by id.
      parameters:
      - description: Advertise id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Advertise
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.AdvertiseInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Advertise'
      security:
      - BearerToken: []
      summary: Update Advertise.
      tags:
      - Advertise
  /login:
    post:
      description: Logging in to get jwt token to access admin or user api by roles.
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login as as user.
      tags:
      - Auth
  /payment:
    get:
      description: Get a list of Payment.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Payment'
            type: array
      summary: Get all Payment.
      tags:
      - Payment
    post:
      description: Creating a new Payment.
      parameters:
      - description: the body to create a new Payment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.PaymentInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
      security:
      - BearerToken: []
      summary: Create New Payment.
      tags:
      - Payment
  /payment-current-user:
    get:
      description: Get All Struct Payment from Curent User.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Payment'
            type: array
      security:
      - BearerToken: []
      summary: Get Struct Payment from Curent User.
      tags:
      - Payment
  /payment/{id}:
    delete:
      description: Delete a Payment by id.
      parameters:
      - description: Payment id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Payment.
      tags:
      - Payment
    get:
      description: Get an Payment by id.
      parameters:
      - description: Payment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
      summary: Get Payment.
      tags:
      - Payment
    patch:
      description: Update Payment by id.
      parameters:
      - description: Payment id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Payment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.PaymentInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
      security:
      - BearerToken: []
      summary: Update Payment.
      tags:
      - Payment
  /register:
    post:
      description: registering a user from public access.
      parameters:
      - description: the body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a user.
      tags:
      - Auth
  /user:
    get:
      description: Get All Struct User from Curent User.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerToken: []
      summary: Get Struct User from Curent User.
      tags:
      - User
  /vocation:
    get:
      description: Get a list of Vocation.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vocation'
            type: array
      summary: Get all Vocation.
      tags:
      - Vocation
    post:
      description: Creating a new Vocation.
      parameters:
      - description: the body to create a new Vocation
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.VocationInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vocation'
      security:
      - BearerToken: []
      summary: Create New Vocation.
      tags:
      - Vocation
  /vocation-current-user:
    get:
      description: Get All Struct Vocation from Curent User.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vocation'
            type: array
      security:
      - BearerToken: []
      summary: Get Struct Vocation from Curent User.
      tags:
      - Vocation
  /vocation/{id}:
    delete:
      description: Delete a Vocation by id.
      parameters:
      - description: Vocation id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Vocation.
      tags:
      - Vocation
    get:
      description: Get an Vocation by id.
      parameters:
      - description: Vocation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vocation'
      summary: Get Vocation.
      tags:
      - Vocation
    patch:
      description: Update Vocation by id.
      parameters:
      - description: Vocation id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Vocation
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.VocationInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vocation'
      security:
      - BearerToken: []
      summary: Update Vocation.
      tags:
      - Vocation
swagger: "2.0"